writing tests is possible now
