+++ test/main-get-tests.spec.js
@@ -1,10 +1,14 @@
-const supertest = require('supertest')(require('../server/routes'))
-const expect = require('chai').expect
+const expect = require('chai').expect;
+const request = require('supertest');
+const myApp = require('../server/routes/index');
+const app = request(myApp);
 
 describe('Server', () => {
     describe('Main GET Route', () => {
         it('checks to see if the user already exists', () => {
-            return supertest
+            app
+            .get('/api')
+            .expect(200)
         })
     })
 })
@@ -58,5 +62,5 @@ describe('Server', () => {
 // })
 
 module.exports = {
-    supertest,
+    app,
 }
